+++ pgruntime/main_test.go
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: 2026 Stefan Majewsky <majewsky@gmx.net>
+// SPDX-License-Identifier: Apache-2.0
+
+package pgruntime_test
+
+import (
+	"testing"
+
+	"go.xyrillian.de/gg/pgruntime"
+)
+
+func TestMain(m *testing.M) {
+	pgruntime.WithTestDB(m, m.Run)
+}